Fearlessly refactor float type to decimal type

+++ BasketUI/BasketLogic/Basket/IBasket.cs
@@ -9,7 +9,7 @@ namespace BasketLogic.Basket
 {
     public interface IBasket
     {
-        float Total { get; }
+        decimal Total { get; }
         void AddItem(BasketItem basketItem);
     }
 }
