+++ BasketUI/BasketLogic/Rules/IRule.cs
@@ -5,6 +5,6 @@ namespace BasketLogic.Rules
 {
     public interface IRule
     {
-        float Apply(IList<BasketItem> basketItems);
+        decimal Apply(IList<BasketItem> basketItems);
     }
 }
